+++ src/main/java/com/mapper/ChatMapper.java
@@ -9,5 +9,5 @@ public interface ChatMapper {
     List<Chat> getChat(int n);
 
     int newChat(Chat chat);
-    List<Chat> getChatByTime(int timestamp);
+    List<Chat> getChatByTime(long timestamp);
 }
